bert_mobile/requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.15.0
accelerate>=0.20.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0

# Tokenization and text processing
sentencepiece>=0.1.99
sacremoses>=0.0.53
nltk>=3.8

# Mobile deployment
onnx>=1.14.0
onnxruntime>=1.16.0

# iOS deployment (CoreML)
coremltools>=7.0

# Android deployment (TensorFlow Lite)
tensorflow>=2.13.0
tf2onnx>=1.15.0

# Model optimization
onnx-simplifier>=0.4.0
optimum>=1.13.0

# Fine-tuning optimizations
peft>=0.6.0
bitsandbytes>=0.41.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0

# Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Web interface (optional)
streamlit>=1.28.0
gradio>=3.50.0

# Hugging Face integration
huggingface_hub>=0.17.0
safetensors>=0.4.0

# Progress tracking
wandb>=0.15.0
tensorboard>=2.14.0

# File handling
jsonlines>=4.0.0
h5py>=3.9.0

# Performance profiling
memory-profiler>=0.61.0
psutil>=5.9.0

# Text processing
regex>=2023.8.8
ftfy>=6.1.1

# Model evaluation
rouge-score>=0.1.2
sacrebleu>=2.3.1
bert-score>=0.3.13

# Deployment utilities
docker>=6.1.0
flask>=2.3.0
gunicorn>=21.2.0
